Updating pip uninstall

+++ .devcontainer/scripts/uninstall/pip/packages.sh
@@ -1,3 +1,5 @@
 #!/usr/bin/env bash
 set -ex
-pip freeze | xargs -I {} pip uninstall -y "{}"
+if command -v pip >/dev/null 2>&1; then
+  pip freeze | xargs -I {} pip uninstall -y "{}"
+fi
